requirements.txt
langchain
langchain-community
langchain-chroma
openai
fastapi
sqlalchemy
typing
psycopg2-binary
uvicorn
transformers
pydantic
torch
dill
pyarrow==14.0.0
pyarrow-hotfix
xxhash
Werkzeug==3.0.0
widgetsnbextension==4.0.11
scikit-learn
scipy
joblib
fsspec[http]==2024.3.1
waitress
python-dotenv
httpx<0.17.0
requests>=2.32.1
pyjwt
python-jose[cryptography]
passlib[bcrypt]
tiktoken
sumy
packaging
atomicwrites
pillow
iniconfig
pluggy
py
toml
gotrue<3.0